Modified BGP Update logic to use the path nexthop value first
If the path has a next hop value set, it should be used in the BGP
Update message first.  This changes to logic to use the check for the
next hop in the order prefix/path->peer config->speaker config.  This
will allow for sending the nexthop as part of the add_prefix message and
allow for overiding the nexthop of the peer if one is set.
Signed-off-by: Alan Quillin <alanquillin@gmail.com>